git-svn-id: file:///mnt/data/Workspace/zie87/game_development/tmp/hare-game-engine/p/trunk@301 67360d68-41fe-4630-88ca-a5c15c9c3197

+++ graphics/ImageSprite.cpp
@@ -43,8 +43,8 @@ namespace hare
     void ImageSprite::renderScene()
     {
         RectF r = rectUV;
-        int origoX = origoPos.x;
-        int origoY = origoPos.y;
+        float origoX = origoPos.x;
+        float origoY = origoPos.y;
         if (!bFaceX){
             r.minX = rectUV.maxX;
             r.maxX = rectUV.minX;
